% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/nlsR2.R
\name{nlsR2}
\alias{nlsR2}
\title{nlsR2}
\usage{
nlsR2(nlsAns, y, p)
}
\arguments{
\item{nlsAns}{A number refers to the residual term.}

\item{y}{A number, it refers to the response vector.}

\item{p}{dataframe. By default it consideres all columns as stressors except the second column stores the system response.}
}
\description{
This function Calculate R^2 and adj R^2 from nls result
}

