% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/plot.R
\name{plot.sgSEM}
\alias{plot.sgSEM}
\title{Plotting of Semi-gSEM}
\usage{
\method{plot}{sgSEM}(x, ..., cutoff = 0.2, width = NULL, height = NULL,
  filename = NULL)
}
\arguments{
\item{x}{The returned list from sgSEM. Plotting uses the first element of this list (res.print) in which the first column of it is response, second column is variable and other columns are corresponding best functional form, r-squared, adj-r-squared, P-value1, P-value2 and P-value3.}

\item{...}{A S3 generic/method consistency.}

\item{cutoff}{A threshold value for adjusted R-squared. Solid lines represent relationship with adjusted R-sqr of greater than the cutoff and dotted lines with less than the cutoff. The default is 0.2.}

\item{width}{A numeric describing the width of the plot output in pixels.}

\item{height}{A numeric describing the height of the plot output in pixels.}

\item{filename}{A character string naming a file to save as an html file.}
}
\value{
An html style plot of pairwise relationship pathway diagram between stressors and repsonses. Arrows show relationships between each variable with given statistical relations along the connection lines.
}
\description{
Plot semi-gSEM result
plot.sgSEM plots a structural equation network model diagram based on best functional form for each selected pairwise variable.
}
\examples{
# Load acrylic data set
data(acrylic)
# Build a semi-gSEM model
ans <- sgSEM(acrylic)
# Plot the network model with adjusted-R-squred of 0.1
plot(ans, cutoff = 0.1)
}

