% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/sgSEMp1.R
\name{sgSEMp1}
\alias{sgSEMp1}
\title{Semi-supervised Generalized Structural Equation Modeling (sgSEM) - Principle 1}
\usage{
sgSEMp1(x, stressor = NULL, response = NULL, nlsInits = data.frame(a1 = 1,
  a2 = 1, a3 = 1))
}
\arguments{
\item{x}{A dataframe. By default the first column of it stores the main stressor, second column stores the system response variable and other columns store different intermediate variables.}

\item{stressor}{A character string of the column name of the main stressor OR a numeric number indexing the column of the main stressor.}

\item{response}{A character string of the column name of the main response OR a numeric number indexing the column of the main response.}

\item{nlsInits}{a data frame of initial vectors for nls. Each column corresponds to a coefficient. The data frame can be generated by the genInit() function. Each row is one initial vecotor. Currently the only nls function included is y = a + b * exp(c * x).}
}
\value{
An object of class sgSEMp1, which is a list of the following items:

\itemize{
\item "table": A matrix. For each row, first column is the response variable, second column is the predictor, the other columns show corresponding summary information: Best functional form, R-squared, adj-R-squared, P-value1, P-value2 and P-value3. The P-values correspond to those of estimators of linear regression coefficients. See details.
\item "bestModels": A matrix. First dimension indicates predictors. The second dimension indicates response variables. The i-jth cell of the matrix stores the name of the best functional form corresponding to the j-th response variable regressed on the i-th predictor.
\item "allModels":  A three dimensional list. The first dimension indicates predictors. The second dimension indicates response variables. Third dimension indicates the fitting results of all 6 functional forms. The i-j-k-th cell of the list stores a "lm" object, corresponding to the j-th response, i-th predictor and the k-th functional form.
}

The object has two added attributes:

\itemize{

\item "attr(res.best, "Step")": A vector. For each variable, it shows in which step it is choosen to be significantly related to response variable.

\item "attr(res.best, "diag.Step")": A matrix. First dimension is predictors; second dimension is response variables. Each cell shows in which step the pairwise relation is being fitted.
}
}
\description{
This function carries out sgSEM principle 1
}
\details{
sgSEMp1 builds a network model of multiple continuous variables. Each pair of variables is tested for sensible paring relation chosen from 6 pre-selected common functional forms in linear regression settings. Adjusted R-squared is used for model selection for every pair.

P-values reported in the "res.print" field of the return list contains the P-values of estimators of linear regression coefficients. The P-values are ordered in the common order of coefficients, i.e. in the order of increasing exponents. For example, in the quadratic functional form y ~ b0 + b1x + b2x^2, the three P-values correspond two those of \\hat{b0}, \\hat{b1} and \\hat{b2}, respectively. If there are less than 3 coefficients to estimate, the extra P-value field is filled with NA's.
}
\examples{
## Load the sample acrylic data set
data(acrylic)

## Run semi-gSEM principle one
ans <- sgSEMp1(acrylic, stressor = "IrradTot", response = "YI", nlsInits = genInit(k = 20))

## Plot the result
plot(ans)

## Plot result with different R-sqr cutoff
plot(ans, cutoff = 0.2)

## Summary
summary(ans)

## Extract relations between IrradTot and IAD2
cf <- path(ans, from = "IrradTot", to = "IAD2")
print(cf)

## Print three components of the result
ans$table
ans$bestModels
ans$allModels
}

