% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/sgSEMp2.R
\name{sgSEMp2}
\alias{sgSEMp2}
\title{Semi-supervised Generalized Structural Equation Modeling (sgSEM) - Principle 2}
\usage{
sgSEMp2(x, stressor = NULL, response = NULL)
}
\arguments{
\item{x}{A dataframe. The first column of it is stressor, second column is response variable and other columns are different predictors.}

\item{stressor}{A character string of the column name of the main stressor OR a numeric number indexing the column of the main stressor.}

\item{response}{A character string of the column name of the main response OR a numeric number indexing the column of the main response.}
}
\value{
A list of the following items:

\itemize{
\item "res.print": A matrix. For each row, first column is the response variable, second column is the predictor, the other columns show corresponding summary information.
\item "res.best": A matrix. First dimension is predictors; second dimension is response variables. Each cell is "lm" object, which is the fitted simple linear regression of the best functional form.
\item "attr(res.best, "Step")": A matrix. First dimension is predictors; second dimension is response variables. Each cell shows in which step the pairwise relation is being fitted.
\item "res.all":  A one dimentional list. Each cell of the list stores a "lm" object, which is the best fitted multiple linear regression for each possible response variable.
\item "attr(res.all, "Step")": A vector. For each variable, it shows in which step it is choosen to be significantly related to response variable.
}
}
\description{
This function carries out sgSEM principle 2
}
\details{
Starting from the system response variable, each variable except for the system stressor is regressed on all other variables except for the system response in an additive multiple regression model, which is reduced by a stepwise selection using stepAIC(). Second, for selected variable, fitted linear regression for 8 pre-seclected funtional forms and pick the best.
}
\examples{
data(acrylic)
ans <- sgSEMp2(acrylic)
ans$res.print
ans$res.best
ans$res.all
attr(ans$res.best, "Step")
attr(ans$res.all, "Step")
}

